Use router.push
