+++ packages/app/src/stores/context.tsx
@@ -56,15 +56,8 @@ export const useCurrentPathname = (initialData?: string): SWRResponse<string, Er
   return useContextSWR('currentPathname', initialData);
 };
 
-// TODO: Consider place https://redmine.weseek.co.jp/issues/108795
-export const useCurrentPageId = (fallbackData?: Nullable<string>): SWRResponse<Nullable<string>, Error> => {
-  const { fallback } = useSWRConfig();
-
-  if (fallbackData !== undefined) {
-    fallback.currentPageId = fallbackData;
-  }
-
-  return useStaticSWR<Nullable<string>, Error>('currentPageId');
+export const useCurrentPageId = (initialData?: Nullable<string>): SWRResponse<Nullable<string>, Error> => {
+  return useStaticSWR<Nullable<string>, Error>('currentPageId', initialData);
 };
 
 export const useIsIdenticalPath = (initialData?: boolean): SWRResponse<boolean, Error> => {
@@ -73,6 +66,10 @@ export const useIsIdenticalPath = (initialData?: boolean): SWRResponse<boolean, 
 
 export const useIsForbidden = (initialData?: boolean): SWRResponse<boolean, Error> => {
   return useContextSWR<boolean, Error>('isForbidden', initialData, { fallbackData: false });
+};
+
+export const useIsNotFound = (initialData?: boolean): SWRResponse<boolean, Error> => {
+  return useContextSWR<boolean, Error>('isNotFound', initialData, { fallbackData: false });
 };
 
 export const useTemplateTagData = (initialData?: Nullable<string>): SWRResponse<Nullable<string>, Error> => {
